+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This changelog's format is based on [keep a changelog 1.0.0](https://keepachange
 ## [Unreleased]
 ### Added
 - population-counter gate
+- performance tests for the bitvector multiplication gate encoder and the ltseq cardinality
+  constraint encoder
 
 ### Changed
 - relaxed the ClauseConsumer interface to accept iterables of literals